+++ custom_components/econet300/entity.py
@@ -84,6 +84,11 @@ class EconetEntity(CoordinatorEntity):
         _LOGGER.debug("Added to HASS: %s", self.entity_description)
         _LOGGER.debug("Coordinator: %s", self.coordinator)
 
+        # Check if the coordinator has a 'data' attributes
+        if "data" not in dir(self.coordinator):
+            _LOGGER.error("Coordinator object does not have a 'data' attribute")
+            return
+
         # Retrieve sysParams and regParams paramsEdits data
         sys_params = self.coordinator.data.get("sysParams", {})
         reg_params = self.coordinator.data.get("regParams", {})
@@ -92,15 +97,10 @@ class EconetEntity(CoordinatorEntity):
         _LOGGER.debug("async_regParams: %s", reg_params)
         _LOGGER.debug("async_paramsEdits: %s", params_edits)
 
-        # Check if the coordinator has a 'data' attributes
-        if "data" not in dir(self.coordinator):
-            _LOGGER.error("Coordinator object does not have a 'data' attribute")
-            return
-
         # Check the available keys in all sources
-        sys_keys = sys_params.keys()
-        reg_keys = reg_params.keys()
-        edit_keys = params_edits.keys()
+        sys_keys = sys_params.keys() if sys_params is not None else []
+        reg_keys = reg_params.keys() if reg_params is not None else []
+        edit_keys = params_edits.keys() if params_edits is not None else []
         _LOGGER.debug("Available keys in sysParams: %s", sys_keys)
         _LOGGER.debug("Available keys in regParams: %s", reg_keys)
         _LOGGER.debug("Available keys in paramsEdits: %s", edit_keys)
@@ -167,7 +167,7 @@ class MixerEntity(EconetEntity):
 
 
 class LambdaEntity(EconetEntity):
-    """Represents EcosterEntity."""
+    """Initialize the LambdaEntity."""
 
     def __init__(
         self,